Rename: pkg/cli -> pkg/clib
