+++ pkg/clib/close.go
@@ -1,4 +1,4 @@
-package cli
+package clib
 
 var closers []func()
