+++ pkg/clib/io.go
@@ -1,4 +1,4 @@
-package cli
+package clib
 
 import (
 	"io"
