+++ pkg/clib/testing/io.go
