+++ pkg/clig/cmd/.snapshots/TestInit-gex_has_already_been_installed--home-go-src-go.example.com-foobar-pkg-foobar-context.go
@@ -1,7 +1,7 @@
 package foobar
 
 import (
-	"github.com/izumin5210/clig/pkg/cli"
+	"github.com/izumin5210/clig/pkg/clib"
 	"github.com/pkg/errors"
 	"github.com/spf13/afero"
 	"github.com/spf13/viper"
@@ -10,13 +10,13 @@ import (
 )
 
 type Ctx struct {
-	WorkingDir cli.Path
-	IO         cli.IO
+	WorkingDir clib.Path
+	IO         clib.IO
 	FS         afero.Fs
 	Viper      *afero.Viper
 	Exec       exec.Interface
 
-	Build  cli.Build
+	Build  clib.Build
 	Config *Config
 }
