+++ pkg/clig/cmd/cmd.go
@@ -5,14 +5,14 @@ import (
 	"github.com/spf13/cobra"
 	"k8s.io/utils/exec"
 
-	"github.com/izumin5210/clig/pkg/cli"
+	"github.com/izumin5210/clig/pkg/clib"
 	"github.com/izumin5210/clig/pkg/clig"
 )
 
-func NewDefaultCligCommand(wd cli.Path, build cli.Build) *cobra.Command {
+func NewDefaultCligCommand(wd clib.Path, build clib.Build) *cobra.Command {
 	return NewCligCommand(&clig.Ctx{
 		WorkingDir: wd,
-		IO:         cli.Stdio(),
+		IO:         clib.Stdio(),
 		FS:         afero.NewOsFs(),
 		Exec:       exec.New(),
 		Build:      build,
@@ -24,11 +24,11 @@ func NewCligCommand(ctx *clig.Ctx) *cobra.Command {
 		Use: ctx.Build.AppName,
 	}
 
-	cli.AddLoggingFlags(cmd)
+	clib.AddLoggingFlags(cmd)
 
 	cmd.AddCommand(
 		newInitCommand(ctx),
-		cli.NewVersionCommand(ctx.IO, ctx.Build),
+		clib.NewVersionCommand(ctx.IO, ctx.Build),
 	)
 
 	return cmd
